+++ tiled/src/test/kotlin/ktx/tiled/MapLayerTest.kt
@@ -1,8 +1,16 @@
 package ktx.tiled
 
+import com.badlogic.gdx.graphics.g2d.TextureRegion
 import com.badlogic.gdx.maps.MapLayer
 import com.badlogic.gdx.maps.MapLayers
-import org.junit.Assert.*
+import com.badlogic.gdx.maps.tiled.TiledMapTileLayer
+import com.badlogic.gdx.maps.tiled.TiledMapTileLayer.Cell
+import com.badlogic.gdx.maps.tiled.tiles.StaticTiledMapTile
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertNull
+import org.junit.Assert.assertTrue
+import org.junit.Assert.fail
 import org.junit.Test
 
 class MapLayerTest {
@@ -81,5 +89,31 @@ class MapLayerTest {
     val actual = MapLayers()
 
     assertFalse(actual.isNotEmpty())
+  }
+
+  @Test
+  fun `should execute action per cell of TiledMapTileLayer`() {
+    val tileLayer =
+      TiledMapTileLayer(2, 2, 32, 32).apply {
+        setCell(0, 0, Cell().setTile(StaticTiledMapTile(TextureRegion())))
+        setCell(1, 1, Cell().setTile(StaticTiledMapTile(TextureRegion())))
+      }
+
+    val visitedCells = mutableSetOf<Pair<Int, Int>>()
+
+    tileLayer.forEachCell { _, cellX, cellY ->
+      visitedCells += cellX to cellY
+    }
+
+    assertEquals(setOf(0 to 0, 1 to 1), visitedCells)
+  }
+
+  @Test
+  fun `should not execute any action for TiledMapTileLayer without cells`() {
+    val tileLayer = TiledMapTileLayer(2, 2, 32, 32)
+
+    tileLayer.forEachCell { _, _, _ ->
+      fail()
+    }
   }
 }